add createTestProject script, rename prev to createTestProjectManifest
