+++ packages/geoprocessing/scripts/testing/types.ts
@@ -1,0 +1,5 @@
+export type TestComponentTypes =
+  | "preprocessor"
+  | "syncGeoprocessor"
+  | "asyncGeoprocessor"
+  | "client";
